rename MapStyleCard
