+++ sites/geohub/src/lib/dashboards/MapStyleCard.svelte
